Implement unshare
Implement the exact functionality of `unshare` instead of just showing
warning

- requireUncached

Firebase library cached the connection for the same url, so when we
release the firebase resources of specific editor, we actually release
the resources of all editors. Thus `unshare` will not work if we are
sharing multiple panes inside the same atom window, but it works if we
are sharing a single pane inside multiple atom windows. The solution is
just requiring uncached Firebase class.

- break for loop

share.remove() mutates @shareStack, we must exit the loop when the
removement is done